// clients/current/captive-portal/src/infrastructure/api/schema.d.ts
export interface components {
    schemas: {
// ...
        PortalBrandingResponse: {
            ssid?: string;
            displayName?: null | string;
            brandPrimary?: string;
            brandPrimaryHover?: string;
            brandSecondary?: string;
            brandAccent?: string;
            textPrimary?: string;
            textSecondary?: string;
            textTertiary?: string;
            textMuted?: string;
            surfaceCard?: string;
            surfaceWhite?: string;
            surfaceBorder?: string;
            buttonPrimary?: string;
            buttonPrimaryHover?: string;
            buttonPrimaryText?: string;
            buttonSecondary?: string;
            buttonSecondaryHover?: string;
            buttonSecondaryText?: string;
            logoUrl?: string;
            logoWhiteUrl?: string;
            connectCardBgUrl?: string;
            bannerOverlayUrl?: null | string;
            faviconUrl?: null | string;
            splashBgUrl?: null | string;
            heading?: null | string;
            subheading?: null | string;
            splashHeading?: null | string;
            buttonText?: null | string;
            termsLinks?: null | string;
            venueLabel?: null | string;
            venueRoute?: null | string;
            /** Format: int32 */
            sortOrder?: number | string;
            authMethods?: string[];
            marketingOptIn?: boolean;
            adsEnabled?: boolean;
            adsReviveServerUrl?: null | string;
            adsReviveZoneId?: null | string;
            adsReviveId?: null | string;
            adsVastUrl?: null | string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}